feat: location in nomissingexternal models linting error
