+++ tests/core/linter/test_builtin.py
@@ -44,7 +44,8 @@ def test_no_missing_external_models(tmp_path, copy_to_temp_path) -> None:
     # Lint the models
     lints = context.lint_models(raise_on_error=False)
     assert len(lints) == 1
+    assert lints[0].violation_range is not None
     assert (
-        "Model 'sushi.customers' depends on unregistered external models: "
-        in lints[0].violation_msg
+        lints[0].violation_msg
+        == """Model '"memory"."sushi"."customers"' depends on unregistered external model '"memory"."raw"."demographics"'. Please register it in the external models file. This can be done by running 'sqlmesh create_external_models'."""
     )
