+++ tests/lsp/test_reference_macro.py
@@ -1,6 +1,6 @@
 from sqlmesh.core.context import Context
 from sqlmesh.lsp.context import LSPContext, ModelTarget
-from sqlmesh.lsp.reference import LSPMacroReference, get_macro_definitions_for_a_path
+from sqlmesh.lsp.reference import MacroReference, get_macro_definitions_for_a_path
 from sqlmesh.lsp.uri import URI
 
 
@@ -24,6 +24,6 @@ def test_macro_references() -> None:
 
     # Check that all references point to the utils.py file
     for ref in macro_references:
-        assert isinstance(ref, LSPMacroReference)
+        assert isinstance(ref, MacroReference)
         assert URI.from_path(ref.path).value.endswith("sushi/macros/utils.py")
         assert ref.target_range is not None
